requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
fastapi==0.100.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
packaging==24.2
    # via pytest
pluggy==1.5.0
    # via pytest
pydantic[dotenv]==2.1.1
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.4.0
    # via pydantic
pydantic-settings==2.0.3
    # via -r requirements.in
pytest==8.3.4
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via -r requirements.in
semantic-version==2.10.0
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.27.0
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.22.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==14.2
    # via uvicorn
